server: move ListenAndUpdateAddrs out of server.go
Release note: None
