+++ pkg/server/listen_and_update_addrs.go
@@ -1,0 +1,53 @@
+// Copyright 2022 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package server
+
+import (
+	"context"
+	"net"
+
+	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/errors"
+)
+
+// ListenError is returned from Start when we fail to start listening on either
+// the main Cockroach port or the HTTP port, so that the CLI can instruct the
+// user on what might have gone wrong.
+type ListenError struct {
+	cause error
+	Addr  string
+}
+
+// Error implements error.
+func (l *ListenError) Error() string { return l.cause.Error() }
+
+// Unwrap is because ListenError is a wrapper.
+func (l *ListenError) Unwrap() error { return l.cause }
+
+// ListenAndUpdateAddrs starts a TCP listener on the specified address
+// then updates the address and advertised address fields based on the
+// actual interface address resolved by the OS during the Listen()
+// call.
+func ListenAndUpdateAddrs(
+	ctx context.Context, addr, advertiseAddr *string, connName string,
+) (net.Listener, error) {
+	ln, err := net.Listen("tcp", *addr)
+	if err != nil {
+		return nil, &ListenError{
+			cause: err,
+			Addr:  *addr,
+		}
+	}
+	if err := base.UpdateAddrs(ctx, addr, advertiseAddr, ln.Addr()); err != nil {
+		return nil, errors.Wrapf(err, "internal error: cannot parse %s listen address", connName)
+	}
+	return ln, nil
+}
